requirements.txt
streamlit==0.85.1
regex==2020.10.15
numpy==1.19.5
nltk==3.5
scikit-learn==0.23.2
scipy==1.6.3
joblib==0.17.0
tensorflow==2.5.0
pytorch==1.8.1
transformers==4.8.2
keras==2.5.0
